LayoutTests/platform/chromium/accessibility/scrollbars-expected.txt

This tests that a scroll area will return scroll bars when asked.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS scrollArea.role is 'AXRole: AXScrollArea'
PASS scrollArea.verticalScrollbar.role is 'AXRole: AXScrollBar'
PASS successfullyParsed is true

TEST COMPLETE
